fix: instructor_profile layout and social icon
