+++ app/(main)/instructor_profile/page.js
@@ -1,0 +1,11 @@
+import InstructorCoursePage from "./_components/CoursePage";
+import ProfileTop from "./_components/ProfileTop";
+
+export default function InstructorProfile() {
+  return (
+    <>
+      <ProfileTop />
+      <InstructorCoursePage />
+    </>
+  );
+}
