+++ package.json
@@ -37,7 +37,6 @@
     "react-dom": "^18",
     "react-dropzone": "^14.3.8",
     "react-hook-form": "^7.54.2",
-    "react-icons": "^5.5.0",
     "react-player": "^2.16.0",
     "sonner": "^2.0.1",
     "swiper": "^11.2.5",
